Renaming from AddEvents to AddEvent
